backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
google-api-python-client==2.127.0
openai==1.24.0
PyMuPDF==1.23.5
statbotics==3.0.0
python-multipart==0.0.7